Spotify API working as expected
